requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
bce-python-sdk==0.9.46
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorlog==6.9.0
cryptography==46.0.1
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
easyocr==1.7.2
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.116.2
fastapi-cli==0.0.12
fastapi-cloud-cli==0.2.0
filelock==3.19.1
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.11.0
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
modelscope==1.30.0
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.108.1
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.2.0
paddlex==3.2.1
pandas==2.3.2
pillow==11.3.0
postgrest==2.20.0
prettytable==3.16.0
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.20.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
sentry-sdk==2.38.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
sympy==1.14.0
tifffile==2025.9.9
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typer==0.18.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
